@@ -11,13 +11,12 @@
         }
     ],
     "require": {
-        "php": ">=5.6",
-        "guzzlehttp/guzzle": "~6.5"
+        "php": ">=7.2",
+        "guzzlehttp/guzzle": "~7"
     },
     "require-dev": {
-        "phpunit/phpunit": "~5",
-        "doctrine/instantiator": "~1.1",
-        "php": ">=5.6"
+        "phpunit/phpunit": "*",
+        "php": ">=7.2"
 
     },
     "autoload": {
